requirements.txt
python-pptx
pdf2image
streamlit==1.41.0 
transformers 
bitsandbytes 
accelerate
gtts
transformers
pillow
torch 
torchvision 
transformers 
pillow
moviepy
timm 
einop
comtypes
google-generativeai